refactor: dummyData.ts 위치 이동 및 관련 import 경로 수정 및 불필요한 파일 삭제
