+++ src/pages/analysis/SelectDiary/index.tsx
@@ -1,9 +1,7 @@
 import S from './style.module.css'
 import NextPageButton from "./components/NextPageButton";
 import DiaryList from "./components/DiaryList";
-import { diaries } from '../../../dummyData';
-// import DiaryItem from './components/DiaryItem';
-// import L from './DiaryLayout.module.css'
+import { diaries } from '../dummyData';
 
 function SelectDiary() {
   return (
@@ -13,7 +11,6 @@ function SelectDiary() {
         <p>감정분석을 하고 싶은 일기를 선택해주세요.</p>
       </section>
       <DiaryList diaries={diaries} />
-      {/* <DiaryItem diary={diaries[1]} /> */}
       <NextPageButton />
     </main>
   )
